backend/requirements.txt
apify-client>=1.6.0
playwright>=1.40.0
python-dotenv>=1.0.0
openai>=1.12.0
google-generativeai>=0.5.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
